+++ android/src/main/java/zhuoyuan/li/fluttershareme/util/ShareFileProvider.java
@@ -3,6 +3,4 @@ package zhuoyuan.li.fluttershareme.util;
 import androidx.core.content.FileProvider;
 
 public class ShareFileProvider extends FileProvider {
-
-
 }